added to not send request if the edit user fields are empty

+++ client/src/components/book-details/BookDetails.jsx
@@ -8,6 +8,7 @@ import StarRating from "../reusables/star-rating/StarRating";
 
 import AuthContext from "../../contexts/authContext";
 import { deleteRatingByBookTitle } from "../../services/ratingService";
+import Spinner from "../reusables/spinner/Spinner";
 
 export default function BookDetails() {
   const [book, setBook] = useState({});
@@ -16,6 +17,7 @@ export default function BookDetails() {
   const navigate = useNavigate();
 
   const [isUnavailable, setIsUnavailabe] = useState(false);
+  const [isLoading, setIsLoading] = useState(false);
 
   const { authState } = useContext(AuthContext);
   const isAuthenticated = authState.isAuthenticated;
@@ -38,11 +40,22 @@ export default function BookDetails() {
   });
 
   const submitDeleteHandler = async () => {
-    await BookService.delete("/books", bookId);
-
-    await deleteRatingByBookTitle(book.title);
-
-    navigate("/books");
+    
+    const isConfirmed = window.confirm("Are you sure you want to delete this book?");
+  
+    if (isConfirmed) {
+      try {
+        
+        const result = await BookService.delete("/books", bookId);
+  
+        if (result) {
+          await deleteRatingByBookTitle(book.title);
+          navigate("/books");
+        }
+      } catch (error) {
+        console.error("Error deleting the book:", error);
+      }
+    } 
   };
 
   const submitGoBackHandler = () => {
@@ -50,6 +63,8 @@ export default function BookDetails() {
   };
 
   const rentHandler = async () => {
+
+    setIsLoading(true)
     try {
       const res = await BookService.updateBook("books", bookId, {
         isRented: true,
@@ -64,22 +79,31 @@ export default function BookDetails() {
     } catch (error) {
       console.log(error);
     }
+
+    const minDuration = 700;
+    setTimeout(() => setIsLoading(false), minDuration);
+    
   };
 
   return (
+    
     <section className="book">
-      <div className="book-container">
+      {isLoading ? (<Spinner/>) : (<><div className="book-container">
+
         <div className="book-img">
+
           <img src={book.img} alt="" />
         </div>
 
         <div className="book-details">
+
           <h1 className="title">{book.title}</h1>
           <label htmlFor="author">Author</label>
           <h3 className="author">{book.author}</h3>
 
           <h2 className="resume">Resume</h2>
           <p>{book.summary}</p>
+
 
           {isAuthenticated && !isOwner.current ? (
             <StarRating totalStars={5} bookId={bookId} bookTitle={book.title} />
@@ -88,30 +112,30 @@ export default function BookDetails() {
             <p>You need to be logged in to see and rate this book.</p>
           )}
         </div>
-      </div>
+      </div><div className="special-buttons">
 
-      <div className="special-buttons">
-        {!isUnavailable && (
-          <button onClick={rentHandler} className="rent-btn">
-            Rent
+          {!isUnavailable && (
+            <button onClick={rentHandler} className="rent-btn">
+              Rent
+            </button>
+          )}
+
+          {isOwner.current && (
+            <Link to={`/books/${bookId}/edit`} className="edit-btn">
+              Edit
+            </Link>
+          )}
+          {isOwner.current && (
+            <button onClick={submitDeleteHandler} className="delete-btn">
+              Delete
+            </button>
+          )}
+
+          <button onClick={submitGoBackHandler} className="go-back-btn">
+            Go Back
           </button>
-        )}
-
-        {isOwner.current && (
-          <Link to={`/books/${bookId}/edit`} className="edit-btn">
-            Edit
-          </Link>
-        )}
-        {isOwner.current && (
-          <button onClick={submitDeleteHandler} className="delete-btn">
-            Delete
-          </button>
-        )}
-
-        <button onClick={submitGoBackHandler} className="go-back-btn">
-          Go Back
-        </button>
-      </div>
+        </div></>)}
+      
     </section>
   );
 }
